# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Enterprise OS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /employees:
    get:
      operationId: listEmployees
      tags: [employees]
      summary: List all employees
      parameters:
        - in: query
          name: department
          schema:
            type: string
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: search
          schema:
            type: string
      responses:
        "200":
          description: Employee list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Employee"
    post:
      operationId: createEmployee
      tags: [employees]
      summary: Create employee
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmployeeInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
  /employees/{id}:
    get:
      operationId: getEmployee
      tags: [employees]
      summary: Get employee by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Employee
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
    patch:
      operationId: updateEmployee
      tags: [employees]
      summary: Update employee
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmployeeUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
    delete:
      operationId: deleteEmployee
      tags: [employees]
      summary: Delete employee
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /departments:
    get:
      operationId: listDepartments
      tags: [departments]
      summary: List departments
      responses:
        "200":
          description: Departments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Department"
    post:
      operationId: createDepartment
      tags: [departments]
      summary: Create department
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DepartmentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Department"
  /departments/{id}:
    patch:
      operationId: updateDepartment
      tags: [departments]
      summary: Update department
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DepartmentUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Department"
    delete:
      operationId: deleteDepartment
      tags: [departments]
      summary: Delete department
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # --- ATTENDANCE ---
...
  /leads:
    get:
      operationId: listLeads
      tags: [leads]
      summary: List CRM leads
      parameters:
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: search
          schema:
            type: string
      responses:
        "200":
          description: Leads
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Lead"
    post:
      operationId: createLead
      tags: [leads]
      summary: Create lead
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
  /leads/{id}:
    get:
      operationId: getLead
      tags: [leads]
      summary: Get lead
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Lead
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
    patch:
      operationId: updateLead
      tags: [leads]
      summary: Update lead
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
    delete:
      operationId: deleteLead
      tags: [leads]
      summary: Delete lead
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /invoices:
    get:
      operationId: listInvoices
      tags: [invoices]
      summary: List invoices
      parameters:
        - in: query
          name: status
          schema:
            type: string
      responses:
        "200":
          description: Invoices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Invoice"
    post:
      operationId: createInvoice
      tags: [invoices]
      summary: Create invoice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvoiceInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
  /invoices/{id}:
    patch:
      operationId: updateInvoice
      tags: [invoices]
      summary: Update invoice
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvoiceUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
    delete:
      operationId: deleteInvoice
      tags: [invoices]
      summary: Delete invoice
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /expenses:
    get:
      operationId: listExpenses
      tags: [expenses]
      summary: List expenses
      parameters:
        - in: query
          name: category
          schema:
            type: string
      responses:
        "200":
          description: Expenses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Expense"
    post:
      operationId: createExpense
      tags: [expenses]
      summary: Create expense
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExpenseInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
  /expenses/{id}:
    patch:
      operationId: updateExpense
      tags: [expenses]
      summary: Update expense
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExpenseUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
    delete:
      operationId: deleteExpense
      tags: [expenses]
      summary: Delete expense
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # --- PROJECTS ---
...
components:
  schemas:
...
    Employee:
      type: object
      required: [id, name, email, department, position, status, joined_date]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        department:
          type: string
        position:
          type: string
        status:
          type: string
        salary:
          type: ["number", "null"]
        joined_date:
          type: string
        avatar:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
    EmployeeInput:
      type: object
      required: [name, email, department, position, status, joined_date]
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        department:
          type: string
        position:
          type: string
        status:
          type: string
        salary:
          type: number
        joined_date:
          type: string
        avatar:
          type: string
        location:
          type: string
    EmployeeUpdate:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        department:
          type: string
        position:
          type: string
        status:
          type: string
        salary:
          type: number
        location:
          type: string
...
    Department:
      type: object
      required: [id, name, head, employee_count]
      properties:
        id:
          type: integer
        name:
          type: string
        head:
          type: ["string", "null"]
        employee_count:
          type: integer
        description:
          type: ["string", "null"]
    DepartmentInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        head:
          type: string
        description:
          type: string
    DepartmentUpdate:
      type: object
      properties:
        name:
          type: string
        head:
          type: string
        description:
          type: string

    # --- ATTENDANCE ---
...
    Lead:
      type: object
      required: [id, name, company, email, status, value, stage, created_at]
      properties:
        id:
          type: integer
        name:
          type: string
        company:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        status:
          type: string
        stage:
          type: string
        value:
          type: number
        assigned_to:
          type: ["string", "null"]
        created_at:
          type: string
    LeadInput:
      type: object
      required: [name, company, email, stage, value]
      properties:
        name:
          type: string
        company:
          type: string
        email:
          type: string
        phone:
          type: string
        status:
          type: string
        stage:
          type: string
        value:
          type: number
        assigned_to:
          type: string
    LeadUpdate:
      type: object
      properties:
        name:
          type: string
        company:
          type: string
        email:
          type: string
        phone:
          type: string
        status:
          type: string
        stage:
          type: string
        value:
          type: number
        assigned_to:
          type: string
...
    Invoice:
      type: object
      required: [id, invoice_number, client, amount, status, issue_date, due_date]
      properties:
        id:
          type: integer
        invoice_number:
          type: string
        client:
          type: string
        amount:
          type: number
        status:
          type: string
        issue_date:
          type: string
        due_date:
          type: string
        description:
          type: ["string", "null"]
    InvoiceInput:
      type: object
      required: [client, amount, issue_date, due_date]
      properties:
        client:
          type: string
        amount:
          type: number
        status:
          type: string
        issue_date:
          type: string
        due_date:
          type: string
        description:
          type: string
    InvoiceUpdate:
      type: object
      properties:
        client:
          type: string
        amount:
          type: number
        status:
          type: string
        due_date:
          type: string
        description:
          type: string
...
    Expense:
      type: object
      required: [id, title, amount, category, date, status]
      properties:
        id:
          type: integer
        title:
          type: string
        amount:
          type: number
        category:
          type: string
        date:
          type: string
        status:
          type: string
        description:
          type: ["string", "null"]
        submitted_by:
          type: ["string", "null"]
    ExpenseInput:
      type: object
      required: [title, amount, category, date]
      properties:
        title:
          type: string
        amount:
          type: number
        category:
          type: string
        date:
          type: string
        status:
          type: string
        description:
          type: string
        submitted_by:
          type: string
    ExpenseUpdate:
      type: object
      properties:
        title:
          type: string
        amount:
          type: number
        category:
          type: string
        status:
          type: string
        description:
          type: string

    # --- PROJECTS ---
...